datasets/generate/requirements.txt
h5py
Pillow
numpy